sketchybar: fix socket timeout on focus change
Still not sure why, but the socket connection will drop when changing
focus to another application. I found this specifically with spotify,
which seems to at some point no longer have a window. When selecting it
via spotlight, the window spawns and we never get a response when
querying all workspaces.

Now there is a 200ms timeout. If we don't get a response quick enough we
just clean up and reconnect, then try again. Works for now.